fix error in 00/ex00 and finish 01/ex03, 01/ex04

+++ Module_00/ex00/megaphone.cpp
@@ -19,7 +19,7 @@ int main(int ac, char **av)
 		std::cout << "* LOUD AND UNBEARABLE FEEDBACK NOISE *" << std::endl;
 	else {
 		// Loop through arguments and append uppercased text
-		for (size_t i = 1; i < ac; i++)
+		for (int i = 1; i < ac; i++)
 			res.append(toUpperCase(av[i]));
 		std::cout << res << std::endl;
 	}
